requirements.txt
django
djangorest
dj-rest-auth
django-cors-headers
djangorestframework-simplejwt
django-filter==2.4.0



https://hannylicious.com/blog/python-requirements-files/
